+++ src/renderer/src/App.tsx
@@ -5,6 +5,7 @@ import { StatusBar } from '@/components/StatusBar'
 import { ConnectionSettings } from '@/components/ConnectionSettings'
 import { SkillsModal } from '@/components/SkillsModal'
 import { SshModal } from '@/components/SshModal'
+import { UsageModal } from '@/components/UsageModal'
 import { HomeView } from '@/views/HomeView'
 import { WorkspaceView } from '@/views/WorkspaceView'
 import { AnalyticsView } from '@/views/AnalyticsView'
@@ -54,6 +55,7 @@ export function App(): JSX.Element {
       <ConnectionSettings />
       <SkillsModal />
       <SshModal />
+      <UsageModal />
     </div>
   )
 }